kerrighed: port KerMM to current kernel APIs
Port the Kerrighed distributed memory subsystem to the current
3.10-based kernel tree.

Update MM, VMA, page-table, highmem, rmap and fault handling APIs,
restore the x86 VDSO import paths for native and compat tasks, and
update KerMM MM state import/export handling.

Harden the COW fault path by initializing the KDDM cow_page argument
and validating __do_fault return states before dereferencing
fault_page.

Keep widened KerMM vm_flags types consistent across madvise, KSM and
MM helpers.

CONFIG_KRG_MM now builds and links successfully and the resulting
kernel boots and joins a two-node Kerrighed cluster.

+++ mm/mmap.c
@@ -3372,7 +3372,7 @@ int special_mapping_vm_ops_krgsyms_unregister(void)
 int install_special_mapping(struct mm_struct *mm,
 			    unsigned long addr, unsigned long len,
 #ifdef CONFIG_KRG_MM
-			    unsigned long long vm_flags, struct page **pages)
+			    unsigned long vm_flags, struct page **pages)
 #else
 			    unsigned long vm_flags, struct page **pages)
 #endif
